Add regression/improvement lists to compare2 view
This patch adds the list of regressed and improved files inline in the
compare2. The lists are hidden by default using HTML5 summary/details.
This is similar to what is done for errors in the main file view.

It is implemented using a new Test_compare.Full module that returns a
list of detailed information for each file.

Only up to 500 regressed and improved files are displayed, to avoid
overloading the page with information.  To view the rest of the
improvements/regressions, if there is more than 500, a separate
paginated view should be created, but this is left for future work.