Add (binary) field to (prover) stanzas
This mostly exposes the existing `binary` field from the `Prover.t`
through the stanzas. If not provided, the `binary` field will be
inferred from the `cmd` as before.

This is useful to define "template" provers that define a shared set of
options in the `cmd` field and can be reused with `inherits` to target
different versions of the same prover with these options.