Fix `same_file` check, reloaded
In #67 the `same_file` check was changed from an (incorrect) comparison
of the device number to a call to `Sys.command`.

There are two issues with this change:

 - This means we are now creating a new process (three new processes,
   actually) for each file when using the `-F` option, which is not very
   efficient.

 - The current code does not use `Filename.quote_command` and hence is
   vulnerable to various code injection shenanigans such as when the
   file is called `$(echo im_in_yo_shell)` (yes that is a valid file
   name).

This patch changes it back to the old way using `Unix.stat`, but
actually checks the inode number in addition to the device number.