Introduce CPU affinity support
This is configured with on the command-line with the new `--cpus`
option, which replaces `-j` and allows to specify a list of cpus or cpu
ranges in taskset format, such as `0-3,7-12,15` (strides are not
supported).

When the `--cpus` option is provided, provers will be run in parallel on
the provided cpus, with at most one prover running on a given cpu at
once.

Note that the `--cpus` setting *only* concerns the prover runs (and some
glue code around one specific prover run): it does *not* otherwise
restrict the cpus used by benchpress itself. It is recommended to use an
external method such as the `taskset` utility to assign to the toplevel
benchpress process a CPU affinity that does not overlap with the prover
CPUs provided in `--cpus`.

When using the `--cpus` setting, users should be aware that *there may
still be other processes on the system using these cpus* (obviously)! It
is thus recommended to use one of the existing techniques to isolate
these CPUs from the rest of the system. I know of two ways to do this:
the `isolcpus` kernel parameter, which is deprecated but is slightly
easier to use, and cpusets, which are not deprecated but harder to use.

To use `isolcpus`, simply set the cpus to use for benchmarking as the
isolated CPUs on the kernel cmdline and reboot (OK, maybe not that
simple for one-shot benchmarking, but fairly easy for a machine that is
mostly used for benchmarking). There is no need to set the CPU affinity
for the `benchpress` binary: it will never be scheduled on the isolated
CPUs, and neither will any other processes (unless manually required
to).

To use cpusets (which is the solution I employed on our benchmark
machine at OCamlPro), you should create a `system` cpuset that contains
only the CPUs that will *NOT* be used by benchpress, and move all
processes to that cpuset (this can be done on a running system, consult
the cpuset documentation). Then, create another cpuset that contains the
CPUs to use for benchpress, including the CPUs to use for the
`benchpress` binary (in practice I use the root cpuset that contains all
the CPUs), and run `benchpress` in that cpuset. You must not forget to
use `taskset` to prevent the `benchpress` binary from using the CPUs
destined for the provers. In practice, I move the shell that I use to
run benchpress to that second cpuset.

This helps with #58