Use single-argument ulimit calls
Some ulimit implementations (at least the one in dash, which is the
/bin/sh implementation used by Debian derivatives) only accept setting
one limit at a time.

```shell
$ dash
$ ulimit -t 20 -m 20
dash: 1: ulimit: too many arguments
```

This causes the limits set by benchpress to be ignored on these
implementations when there are multiple (e.g. time and memory) limits.
This patch changes the behavior to build a sequence of unrelated calls
to `ulimit` (so `ulimit -t 20 ; ulimit -m 20 ; my_prover` instead of
`ulimit -t 20 -m 20 ; my_prover`) for compatibility with these
implementations.