Cache global metadata in separate meta.sqlite3 database
This makes it faster to access the prover metadata for all runs in the
`compare2` view because we don't have to open a bajillion different
databases (goes from several minutes to a few seconds for me when
restarting the server).

Note that the database cache is not currently garbage collected and
might need to be manually cleaned periodically.